pyproject.toml
[tool.poetry]
name = "desafio"
version = "0.1.0"
description = ""
authors = ["Mateus Filpo <mateusfilpo27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
